Create player_skye.py
After going through the project description, I realized that there was no possibility of a game tie. We need to beat the basic AI opponent. And if there is a chance to win, the faster one will win. So I try to make the average turn for AI as fast as I can.